+++ tasks/question_answering.py
@@ -10,10 +10,15 @@ from transformers import (
     default_data_collator,
     BitsAndBytesConfig,
 )
-from nyuntam_adapt.core.custom_model import prepare_custom_model_support
+from .custom_model import prepare_custom_model_support
 from nyuntam_adapt.core.base_task import BaseTask
-from nyuntam_adapt.utils.task_utils import prepare_model_for_kbit_training
-from nyuntam_adapt.utils.task_utils import ModelLoadingError
+from nyuntam_adapt.utils import prepare_model_for_kbit_training
+
+
+class ModelLoadingError(RuntimeError):
+    """Exception for custom model loading errors."""
+
+    pass
 
 
 class QuestionAnswering(BaseTask):
@@ -217,9 +222,7 @@ class QuestionAnswering(BaseTask):
                         **self.model_args,
                     )
             except Exception as e:
-                raise ModelLoadingError(
-                    f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
-                ) from e
+                raise ModelLoadingError(f"Following Error Happened : {e}") from e
 
         if use_bnb:
             model = prepare_model_for_kbit_training(model, self.gradient_checkpointing)
