--- tasks/question_answering_hf/__init__.py
@@ -1,1 +1,0 @@
-from .task import *
