+++ tasks/translation.py
@@ -9,9 +9,14 @@ from transformers import (
     BitsAndBytesConfig,
 )
 from nyuntam_adapt.core.base_task import BaseTask
-from nyuntam_adapt.utils.task_utils import prepare_model_for_kbit_training
-from nyuntam_adapt.core.custom_model import prepare_custom_model_support
-from nyuntam_adapt.utils.task_utils import ModelLoadingError
+from nyuntam_adapt.utils import prepare_model_for_kbit_training
+from .custom_model import prepare_custom_model_support
+
+
+class ModelLoadingError(RuntimeError):
+    """Exception for custom model loading errors."""
+
+    pass
 
 
 class Translation(BaseTask):
@@ -106,9 +111,7 @@ class Translation(BaseTask):
                         use_flash_attention_2=self.flash_attention,
                     )
             except Exception as e:
-                raise ModelLoadingError(
-                    f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
-                ) from e
+                raise ModelLoadingError(f"Following Error Happened : {e}") from e
 
         if use_bnb:
             model = prepare_model_for_kbit_training(model, self.gradient_checkpointing)
